feat:ì Course 엔티티 수정 ë및 ㅊãcreateLesson ìì생성

+++ src/main/java/likelion/tupl/dto/LessonDto.java
@@ -27,5 +27,5 @@ public class LessonDto {
     private String studyContent; // 오늘 나간 진도
     private Integer currentLessonTime;// 현재 회차
     private Integer totalLessonTime; // 전체 회차
-    private Integer courseId; // 연결된 Course의 ID(FK)
+    private Long courseId; // 연결된 Course의 ID(FK)
 }
